Extend unmarshalling and start with Juce independent API
Based on discussions, there is the plan to use NanoOCP in future
without the need of Juce. For this the unmarshal functionality
is extended to use std::vector instead of Juce::MemoryBlock.

For backward compatibility, the API with Juce is kept until
a breaking change version is released.

Also it is necessary to unmarshal more OCA packages as it is done
before.